requirements.txt
beautifulsoup4==4.5.3
nltk==3.2.2
lxml==3.8.0
numpy==1.12.1
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
snowballstemmer==1.2.1
torchfile==0.0.2
unicodecsv==0.14.1
torchwordemb==0.0.7
